requirements.txt
google-genai
python-dotenv
openai
numpy
fastapi
uvicorn[standard]
redis